Clean up locations layer
